Restructuring to match idiomatic go layout
This makes the package much easier to import and use.

There should be no functional changes in this commit.
