+++ src/bot/bot/cmd_test.go
