+++ src/bot/bot/parser.go
