+++ src/bot/bot/parser_test.go
