+++ src/bot/irc/irc_test.go
